srv_reader/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
fastapi==0.92.0
uvicorn>=0.15.0,<0.16.0